requirements.txt
robotframework~=4.1.3
selenium~=4.1.3
requests~=2.27.1
smart-assertions~=1.0.2
py-jsonic~=0.0.2